Make website working again
FossilOrigin-Name: 4fa10d3079b3751d46fdfb71dad1a883c675fa459266fe96bcbb51e0e3cc93c2

+++ src/www/admin/web/new.php
@@ -12,7 +12,7 @@ require_once __DIR__ . '/_inc.php';
 $csrf_key = 'web_page_new';
 
 $form->runIf('create', function () {
-	$page = Page::create((int) qg('type'), (int) qg('parent') ?: null, f('title'), Page::STATUS_DRAFT);
+	$page = Page::create((int) qg('type'), qg('parent') ?: null, f('title'), Page::STATUS_DRAFT);
 	$page->save();
 	Utils::redirect(ADMIN_URL . 'web/edit.php?new&id=' . $page->id());
 }, $csrf_key);
